requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
distlib==0.3.9
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.2
django-cache-url==3.4.5
django-crispy-forms==2.4
environs==14.1.1
filelock==3.18.0
gunicorn==23.0.0
marshmallow==4.0.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
virtualenv==20.30.0
whitenoise==6.9.0
